+++ plugin/src/main/java/nl/esciencecenter/e3dchem/knime/molviewer/InvalidFormatException.java
@@ -1,0 +1,10 @@
+package nl.esciencecenter.e3dchem.knime.molviewer;
+
+public class InvalidFormatException extends Exception {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = -6860780420131282998L;
+
+}