+++ server/src/main/java/nl/esciencecenter/e3dchem/knime/molviewer/server/resources/MoleculeResource.java
@@ -14,7 +14,7 @@ public class MoleculeResource {
 	private List<Molecule> molecules;
 
 	public MoleculeResource() {
-		this.molecules = new ArrayList<Molecule>();
+		this.molecules = new ArrayList<>();
 	}
 
 	public MoleculeResource(List<Molecule> molecules) {